refactor(BridgeGame): models 폴더로 이동
